+++ Program.cs
@@ -17,6 +17,7 @@ internal static class Program
 
         app.Add<TgCommands>();
         app.Add<ChatCommands>("chat");
+        app.Add<ChannelCommands>("channel");
         app.Add<LinkCommands>("link");
         app.Add<MessageCommands>("message");